Introduction

Introduction

ElTopo.jl is a interface to a C++ library eltopo. ElTopo the C++ library is a package for tracking dynamic surfaces represented as triangle meshes in 3D. It robustly handles topology changes such as merging and pinching off, while adaptively maintaining a tangle-free, high-quality triangulation. Suitable to simulate time dependant boundary problems, such as droplet behaviour uppon external field, interface dynamics of liquids in a container, etc.

Currently, the wrapper works only on Linux (Help wanted! See ElTopoBuilder and Travis ) and does not expose internal triangular mesh data structure methods. Nevertheless, it provides the mesh stabilization method with expossed parameters to control it. And together with SurfaceTopology.jl should be useful for soft surface dynamics calculations.

The API

The wrapper focuses on the high-level functionality provided in ElTopo. There are three important functions in it SurfTrack::improve_mesh, SurfTrack::topology_changes and SurfTrack::integrate. The first two methods are combined under the name stabilize whereas the latter one, which modifies tangential velocities to avoid vertex collisions, is not wrapped [1].

ElTopo.SurfTrack — Type.

SurfTrack([...])

Constructs a SurfTrack instance which holds parameters for ElTopo surface tracker stabilization.

source

ElTopo.stabilize — Method.

stabilize(msh::HomogenousMesh,p::SurfTrack)

Computes stabilized mesh for a given initial triangulation msh with stabilization parameters p.

source

Let's first load ElTopo library and load the triangulation to stabilize. In this case, a sphere made out of two subdivisions of an icosahedron.

using ElTopo
include("sphere.jl")
msh = unitsphere(2)

HomogenousMesh(
    faces: 320xGeometryTypes.Face{3,Int64},     vertices: 162xGeometryTypes.Point{3,Float64}, )

Berore executing stabilization method we need to initialize parameters. See docstrings for fields and also article.

scale = 0.2
par = SurfTrack(
    min_edge_length = 0.7*scale,
    max_edge_length = 1.5*scale,
    max_volume_change = 0.1*scale^3,
    merge_proximity_epsilon = 0.5*scale,
)

SurfTrack
  proximity_epsilon: Float64 0.0001
  friction_coefficient: Float64 0.0
  min_triangle_area: Float64 1.0e-7
  improve_collision_epsilon: Float64 2.0e-6
  use_fraction: Bool false
  min_edge_length: Float64 0.13999999999999999
  max_edge_length: Float64 0.30000000000000004
  max_volume_change: Float64 0.0008000000000000003
  min_triangle_angle: Float64 0.0
  max_triangle_angle: Float64 180.0
  use_curvature_when_splitting: Bool false
  use_curvature_when_collapsing: Bool false
  min_curvature_multiplier: Float64 1.0
  max_curvature_multiplier: Float64 1.0
  allow_vertex_movement: Bool false
  edge_flip_min_length_change: Float64 0.05
  merge_proximity_epsilon: Float64 0.1
  collision_safety: Bool true
  allow_topology_changes: Bool false
  allow_non_manifold: Bool false
  perform_improvement: Bool true

Then putting parameters and mesh in we can execute stabilization:

newmsh = stabilize(msh,par)

HomogenousMesh(
    faces: 800xGeometryTypes.Face{3,UInt64},     vertices: 402xGeometryTypes.Point{3,Float64}, )

[1]
I had issues with using Cxx to represent SurfTrack instance in julia. Thus the best I was able to do was to use a single icxx block in julia function to do all the stuff. In spite of that, I think that covers most of the use cases.